Refactored `source/io/iomap_otbm.cpp` into modular files in `source/io/otbm/`: - `map.cpp` (Main logic) - `item.cpp` (Item methods) - `spawn.cpp` (Spawn loading/saving) - `house.cpp` (House loading/saving) - `waypoint.cpp` (Waypoint loading/saving)
Relocated `source/ui/pngfiles.{cpp,h}` to `source/resources/png_data.{cpp,h}` to separate data from UI logic.
Updated `CMakeLists.txt` and includes accordingly.

Also fixed a build issue in `source/app/application.cpp` related to `SetAppearance` compatibility with older wxWidgets versions.
