+++ source/io/otbm/house.cpp
@@ -1,0 +1,166 @@
+//////////////////////////////////////////////////////////////////////
+// This file is part of Remere's Map Editor
+//////////////////////////////////////////////////////////////////////
+// Remere's Map Editor is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Remere's Map Editor is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+//////////////////////////////////////////////////////////////////////
+
+#include "app/main.h"
+
+#include <wx/wfstream.h>
+#include <wx/tarstrm.h>
+#include <wx/zstream.h>
+#include <wx/mstream.h>
+#include <wx/datstrm.h>
+
+#include <format>
+#include <fstream>
+#include <vector>
+
+#include "app/settings.h"
+#include "ui/gui.h"
+#include "ui/dialog_util.h" // Loadbar
+
+#include "game/creatures.h"
+#include "game/creature.h"
+#include "map/map.h"
+#include "map/tile.h"
+#include "game/item.h"
+#include "game/complexitem.h"
+#include "game/town.h"
+#include "brushes/wall/wall_brush.h"
+
+#include "io/otbm/map.h"
+#include <spdlog/spdlog.h>
+
+using attribute_t = uint8_t;
+using flags_t = uint32_t;
+bool IOMapOTBM::loadHouses(Map& map, const FileName& dir) {
+	std::string fn = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvUTF8));
+	fn += map.housefile;
+
+	FileName filename(wxstr(fn));
+	if (!filename.FileExists()) {
+		warnings.push_back("IOMapOTBM::loadHouses: File not found.");
+		return false;
+	}
+
+	// has to be declared again as encoding-specific characters break loading there
+	std::string encoded_path = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvWhateverWorks));
+	encoded_path += map.housefile;
+	pugi::xml_document doc;
+	pugi::xml_parse_result result = doc.load_file(encoded_path.c_str());
+	if (!result) {
+		warnings.push_back("IOMapOTBM::loadHouses: File loading error.");
+		return false;
+	}
+	return loadHouses(map, doc);
+}
+
+bool IOMapOTBM::loadHouses(Map& map, pugi::xml_document& doc) {
+	pugi::xml_node node = doc.child("houses");
+	if (!node) {
+		warnings.push_back("IOMapOTBM::loadHouses: Invalid rootheader.");
+		return false;
+	}
+
+	pugi::xml_attribute attribute;
+	for (pugi::xml_node houseNode = node.first_child(); houseNode; houseNode = houseNode.next_sibling()) {
+		if (as_lower_str(houseNode.name()) != "house") {
+			continue;
+		}
+
+		House* house = nullptr;
+		if ((attribute = houseNode.attribute("houseid"))) {
+			house = map.houses.getHouse(attribute.as_uint());
+			if (!house) {
+				break;
+			}
+		}
+
+		if ((attribute = houseNode.attribute("name"))) {
+			house->name = attribute.as_string();
+		} else {
+			house->name = "House #" + std::to_string(house->getID());
+		}
+
+		Position exitPosition(
+			houseNode.attribute("entryx").as_int(),
+			houseNode.attribute("entryy").as_int(),
+			houseNode.attribute("entryz").as_int()
+		);
+		if (exitPosition.x != 0 && exitPosition.y != 0 && exitPosition.z != 0) {
+			house->setExit(exitPosition);
+		}
+
+		if ((attribute = houseNode.attribute("rent"))) {
+			house->rent = attribute.as_int();
+		}
+
+		if ((attribute = houseNode.attribute("guildhall"))) {
+			house->guildhall = attribute.as_bool();
+		}
+
+		if ((attribute = houseNode.attribute("townid"))) {
+			house->townid = attribute.as_uint();
+		} else {
+			warning("House %d has no town! House was removed.", house->getID());
+			map.houses.removeHouse(house);
+		}
+	}
+	return true;
+}
+
+bool IOMapOTBM::saveHouses(Map& map, const FileName& dir) {
+	wxString filepath = dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME);
+	filepath += wxString(map.housefile.c_str(), wxConvUTF8);
+
+	// Create the XML file
+	pugi::xml_document doc;
+	if (saveHouses(map, doc)) {
+		return doc.save_file(filepath.wc_str(), "\t", pugi::format_default, pugi::encoding_utf8);
+	}
+	return false;
+}
+
+bool IOMapOTBM::saveHouses(Map& map, pugi::xml_document& doc) {
+	pugi::xml_node decl = doc.prepend_child(pugi::node_declaration);
+	if (!decl) {
+		return false;
+	}
+
+	decl.append_attribute("version") = "1.0";
+
+	pugi::xml_node houseNodes = doc.append_child("houses");
+	for (const auto& houseEntry : map.houses) {
+		const House* house = houseEntry.second;
+		pugi::xml_node houseNode = houseNodes.append_child("house");
+
+		houseNode.append_attribute("name") = house->name.c_str();
+		houseNode.append_attribute("houseid") = house->getID();
+
+		const Position& exitPosition = house->getExit();
+		houseNode.append_attribute("entryx") = exitPosition.x;
+		houseNode.append_attribute("entryy") = exitPosition.y;
+		houseNode.append_attribute("entryz") = exitPosition.z;
+
+		houseNode.append_attribute("rent") = house->rent;
+		if (house->guildhall) {
+			houseNode.append_attribute("guildhall") = true;
+		}
+
+		houseNode.append_attribute("townid") = house->townid;
+		houseNode.append_attribute("size") = static_cast<int32_t>(house->size());
+	}
+	return true;
+}
