+++ source/io/otbm/map.cpp
@@ -40,489 +40,11 @@
 #include "game/town.h"
 #include "brushes/wall/wall_brush.h"
 
-#include "io/iomap_otbm.h"
+#include "io/otbm/map.h"
 #include <spdlog/spdlog.h>
 
 using attribute_t = uint8_t;
 using flags_t = uint32_t;
-
-// H4X
-void reform(Map* map, Tile* tile, Item* item) {
-	/*
-	int aid = item->getActionID();
-	int id = item->getID();
-	int uid = item->getUniqueID();
-
-	if(item->isDoor()) {
-		item->eraseAttribute("aid");
-		item->setAttribute("keyid", aid);
-	}
-
-	if((item->isDoor()) && tile && tile->getHouseID()) {
-		Door* self = static_cast<Door*>(item);
-		House* house = map->houses.getHouse(tile->getHouseID());
-		self->setDoorID(house->getEmptyDoorID());
-	}
-	*/
-}
-
-// ============================================================================
-// Item
-
-Item* Item::Create_OTBM(const IOMap& maphandle, BinaryNode* stream) {
-	uint16_t _id;
-	if (!stream->getU16(_id)) {
-		return nullptr;
-	}
-
-	uint8_t _count = 0;
-
-	const ItemType& iType = g_items[_id];
-	if (maphandle.version.otbm == MAP_OTBM_1) {
-		if (iType.stackable || iType.isSplash() || iType.isFluidContainer()) {
-			stream->getU8(_count);
-		}
-	}
-	return Item::Create(_id, _count);
-}
-
-bool Item::readItemAttribute_OTBM(const IOMap& maphandle, OTBM_ItemAttribute attr, BinaryNode* stream) {
-	switch (attr) {
-		case OTBM_ATTR_COUNT: {
-			uint8_t subtype;
-			if (!stream->getU8(subtype)) {
-				return false;
-			}
-			setSubtype(subtype);
-			break;
-		}
-		case OTBM_ATTR_ACTION_ID: {
-			uint16_t aid;
-			if (!stream->getU16(aid)) {
-				return false;
-			}
-			setActionID(aid);
-			break;
-		}
-		case OTBM_ATTR_UNIQUE_ID: {
-			uint16_t uid;
-			if (!stream->getU16(uid)) {
-				return false;
-			}
-			setUniqueID(uid);
-			break;
-		}
-		case OTBM_ATTR_CHARGES: {
-			uint16_t charges;
-			if (!stream->getU16(charges)) {
-				return false;
-			}
-			setSubtype(charges);
-			break;
-		}
-		case OTBM_ATTR_TEXT: {
-			std::string text;
-			if (!stream->getString(text)) {
-				return false;
-			}
-			setText(text);
-			break;
-		}
-		case OTBM_ATTR_DESC: {
-			std::string text;
-			if (!stream->getString(text)) {
-				return false;
-			}
-			setDescription(text);
-			break;
-		}
-		case OTBM_ATTR_RUNE_CHARGES: {
-			uint8_t subtype;
-			if (!stream->getU8(subtype)) {
-				return false;
-			}
-			setSubtype(subtype);
-			break;
-		}
-		case OTBM_ATTR_TIER: {
-			uint8_t tier;
-			if (!stream->getU8(tier)) {
-				return false;
-			}
-			setTier(static_cast<uint16_t>(tier));
-			break;
-		}
-
-		// The following *should* be handled in the derived classes
-		// However, we still need to handle them here since otherwise things
-		// will break horribly
-		case OTBM_ATTR_DEPOT_ID:
-			return stream->skip(2);
-		case OTBM_ATTR_HOUSEDOORID:
-			return stream->skip(1);
-		case OTBM_ATTR_TELE_DEST:
-			return stream->skip(5);
-		case OTBM_ATTR_PODIUMOUTFIT:
-			return stream->skip(15);
-		default:
-			return false;
-	}
-	return true;
-}
-
-bool Item::unserializeAttributes_OTBM(const IOMap& maphandle, BinaryNode* stream) {
-	uint8_t attribute;
-	while (stream->getU8(attribute)) {
-		if (attribute == OTBM_ATTR_ATTRIBUTE_MAP) {
-			if (!ItemAttributes::unserializeAttributeMap(maphandle, stream)) {
-				return false;
-			}
-		} else if (!readItemAttribute_OTBM(maphandle, static_cast<OTBM_ItemAttribute>(attribute), stream)) {
-			return false;
-		}
-	}
-	return true;
-}
-
-bool Item::unserializeItemNode_OTBM(const IOMap& maphandle, BinaryNode* node) {
-	return unserializeAttributes_OTBM(maphandle, node);
-}
-
-void Item::serializeItemAttributes_OTBM(const IOMap& maphandle, NodeFileWriteHandle& stream) const {
-	if (maphandle.version.otbm >= MAP_OTBM_2) {
-		const ItemType& iType = g_items[id];
-		if (iType.stackable || iType.isSplash() || iType.isFluidContainer()) {
-			stream.addU8(OTBM_ATTR_COUNT);
-			stream.addU8(getSubtype());
-		}
-	}
-
-	if (maphandle.version.otbm >= MAP_OTBM_4) {
-		if (attributes && !attributes->empty()) {
-			stream.addU8(OTBM_ATTR_ATTRIBUTE_MAP);
-			serializeAttributeMap(maphandle, stream);
-		}
-	} else {
-		if (g_items.MinorVersion >= CLIENT_VERSION_820 && isCharged()) {
-			stream.addU8(OTBM_ATTR_CHARGES);
-			stream.addU16(getSubtype());
-		}
-
-		uint16_t actionId = getActionID();
-		if (actionId > 0) {
-			stream.addU8(OTBM_ATTR_ACTION_ID);
-			stream.addU16(actionId);
-		}
-
-		uint16_t uniqueId = getUniqueID();
-		if (uniqueId > 0) {
-			stream.addU8(OTBM_ATTR_UNIQUE_ID);
-			stream.addU16(uniqueId);
-		}
-
-		std::string text(getText());
-		if (!text.empty()) {
-			stream.addU8(OTBM_ATTR_TEXT);
-			stream.addString(text);
-		}
-
-		std::string description(getDescription());
-		if (!description.empty()) {
-			stream.addU8(OTBM_ATTR_DESC);
-			stream.addString(description);
-		}
-
-		uint16_t tier = getTier();
-		if (tier > 0) {
-			stream.addU8(OTBM_ATTR_TIER);
-			stream.addU8(static_cast<uint8_t>(tier));
-		}
-	}
-}
-
-void Item::serializeItemCompact_OTBM(const IOMap& maphandle, NodeFileWriteHandle& stream) const {
-	stream.addU16(id);
-
-	/* This is impossible
-	const ItemType& iType = g_items[id];
-
-	if(iType.stackable || iType.isSplash() || iType.isFluidContainer()){
-		stream.addU8(getSubtype());
-	}
-	*/
-}
-
-bool Item::serializeItemNode_OTBM(const IOMap& maphandle, NodeFileWriteHandle& file) const {
-	file.addNode(OTBM_ITEM);
-	file.addU16(id);
-	if (maphandle.version.otbm == MAP_OTBM_1) {
-		const ItemType& iType = g_items[id];
-		if (iType.stackable || iType.isSplash() || iType.isFluidContainer()) {
-			file.addU8(getSubtype());
-		}
-	}
-	serializeItemAttributes_OTBM(maphandle, file);
-	file.endNode();
-	return true;
-}
-
-// ============================================================================
-// Teleport
-
-bool Teleport::readItemAttribute_OTBM(const IOMap& maphandle, OTBM_ItemAttribute attribute, BinaryNode* stream) {
-	if (OTBM_ATTR_TELE_DEST == attribute) {
-		uint16_t x, y;
-		uint8_t z;
-		if (!stream->getU16(x) || !stream->getU16(y) || !stream->getU8(z)) {
-			return false;
-		}
-		destination = Position(x, y, z);
-		return true;
-	} else {
-		return Item::readItemAttribute_OTBM(maphandle, attribute, stream);
-	}
-}
-
-void Teleport::serializeItemAttributes_OTBM(const IOMap& maphandle, NodeFileWriteHandle& stream) const {
-	Item::serializeItemAttributes_OTBM(maphandle, stream);
-
-	stream.addByte(OTBM_ATTR_TELE_DEST);
-	stream.addU16(destination.x);
-	stream.addU16(destination.y);
-	stream.addU8(destination.z);
-}
-
-// ============================================================================
-// Door
-
-bool Door::readItemAttribute_OTBM(const IOMap& maphandle, OTBM_ItemAttribute attribute, BinaryNode* stream) {
-	if (OTBM_ATTR_HOUSEDOORID == attribute) {
-		uint8_t id = 0;
-		if (!stream->getU8(id)) {
-			return false;
-		}
-		doorId = id;
-		return true;
-	} else {
-		return Item::readItemAttribute_OTBM(maphandle, attribute, stream);
-	}
-}
-
-void Door::serializeItemAttributes_OTBM(const IOMap& maphandle, NodeFileWriteHandle& stream) const {
-	Item::serializeItemAttributes_OTBM(maphandle, stream);
-	if (doorId) {
-		stream.addByte(OTBM_ATTR_HOUSEDOORID);
-		stream.addU8(doorId);
-	}
-}
-
-DoorType Door::getDoorType() const {
-	WallBrush* wb = getWallBrush();
-	if (!wb) {
-		return WALL_UNDEFINED;
-	}
-
-	return wb->getDoorTypeFromID(id);
-}
-
-bool Door::isRealDoor() const {
-	const DoorType& dt = getDoorType();
-	// doors with no wallbrush will appear as WALL_UNDEFINED
-	// this is for compatibility
-	return dt == WALL_UNDEFINED || dt == WALL_DOOR_NORMAL || dt == WALL_DOOR_LOCKED || dt == WALL_DOOR_QUEST || dt == WALL_DOOR_MAGIC || dt == WALL_DOOR_NORMAL_ALT;
-}
-
-uint8_t Door::getDoorID() const {
-	return isRealDoor() ? doorId : 0;
-}
-
-void Door::setDoorID(uint8_t id) {
-	doorId = isRealDoor() ? id : 0;
-}
-
-// ============================================================================
-// Depots
-
-bool Depot::readItemAttribute_OTBM(const IOMap& maphandle, OTBM_ItemAttribute attribute, BinaryNode* stream) {
-	if (OTBM_ATTR_DEPOT_ID == attribute) {
-		uint16_t id = 0;
-		if (!stream->getU16(id)) {
-			return false;
-		}
-		depotId = id;
-		return true;
-	} else {
-		return Item::readItemAttribute_OTBM(maphandle, attribute, stream);
-	}
-}
-
-void Depot::serializeItemAttributes_OTBM(const IOMap& maphandle, NodeFileWriteHandle& stream) const {
-	Item::serializeItemAttributes_OTBM(maphandle, stream);
-	if (depotId) {
-		stream.addByte(OTBM_ATTR_DEPOT_ID);
-		stream.addU16(depotId);
-	}
-}
-
-// ============================================================================
-// Container
-
-bool Container::unserializeItemNode_OTBM(const IOMap& maphandle, BinaryNode* node) {
-	if (!Item::unserializeAttributes_OTBM(maphandle, node)) {
-		return false;
-	}
-
-	BinaryNode* child = node->getChild();
-	if (child) {
-		do {
-			uint8_t type;
-			if (!child->getByte(type)) {
-				return false;
-			}
-
-			if (type != OTBM_ITEM) {
-				return false;
-			}
-
-			Item* item = Item::Create_OTBM(maphandle, child);
-			if (!item) {
-				return false;
-			}
-
-			if (!item->unserializeItemNode_OTBM(maphandle, child)) {
-				delete item;
-				return false;
-			}
-
-			contents.push_back(item);
-		} while (child->advance());
-	}
-	return true;
-}
-
-bool Container::serializeItemNode_OTBM(const IOMap& maphandle, NodeFileWriteHandle& file) const {
-	file.addNode(OTBM_ITEM);
-	file.addU16(id);
-	if (maphandle.version.otbm == MAP_OTBM_1) {
-		// In the ludicrous event that an item is a container AND stackable, we have to do this. :p
-		const ItemType& iType = g_items[id];
-		if (iType.stackable || iType.isSplash() || iType.isFluidContainer()) {
-			file.addU8(getSubtype());
-		}
-	}
-
-	serializeItemAttributes_OTBM(maphandle, file);
-	for (Item* item : contents) {
-		item->serializeItemNode_OTBM(maphandle, file);
-	}
-
-	file.endNode();
-	return true;
-}
-
-// ============================================================================
-// Podium
-
-bool Podium::readItemAttribute_OTBM(const IOMap& maphandle, OTBM_ItemAttribute attribute, BinaryNode* stream) {
-	if (OTBM_ATTR_PODIUMOUTFIT == attribute) {
-		uint8_t flags;
-		uint8_t direction;
-
-		uint16_t lookType;
-		uint8_t lookHead;
-		uint8_t lookBody;
-		uint8_t lookLegs;
-		uint8_t lookFeet;
-		uint8_t lookAddon;
-
-		uint16_t lookMount;
-		uint8_t lookMountHead;
-		uint8_t lookMountBody;
-		uint8_t lookMountLegs;
-		uint8_t lookMountFeet;
-
-		if (
-			// podium settings
-			stream->getU8(flags) && stream->getU8(direction) &&
-
-			// outfit
-			stream->getU16(lookType) && stream->getU8(lookHead) && stream->getU8(lookBody) && stream->getU8(lookLegs) && stream->getU8(lookFeet) && stream->getU8(lookAddon) &&
-
-			// mount
-			stream->getU16(lookMount) && stream->getU8(lookMountHead) && stream->getU8(lookMountBody) && stream->getU8(lookMountLegs) && stream->getU8(lookMountFeet)
-		) { //"if" condition ends here
-
-			setShowOutfit((flags & PODIUM_SHOW_OUTFIT) != 0);
-			setShowMount((flags & PODIUM_SHOW_MOUNT) != 0);
-			setShowPlatform((flags & PODIUM_SHOW_PLATFORM) != 0);
-
-			setDirection(static_cast<Direction>(direction));
-
-			struct Outfit newOutfit = Outfit();
-			newOutfit.lookType = static_cast<int>(lookType);
-			newOutfit.lookHead = static_cast<int>(lookHead);
-			newOutfit.lookBody = static_cast<int>(lookBody);
-			newOutfit.lookLegs = static_cast<int>(lookLegs);
-			newOutfit.lookFeet = static_cast<int>(lookFeet);
-			newOutfit.lookAddon = static_cast<int>(lookAddon);
-			newOutfit.lookMount = static_cast<int>(lookMount);
-			newOutfit.lookMountHead = static_cast<int>(lookMountHead);
-			newOutfit.lookMountBody = static_cast<int>(lookMountBody);
-			newOutfit.lookMountLegs = static_cast<int>(lookMountLegs);
-			newOutfit.lookMountFeet = static_cast<int>(lookMountFeet);
-			setOutfit(newOutfit);
-			return true;
-		}
-		return false;
-	} else {
-		return Item::readItemAttribute_OTBM(maphandle, attribute, stream);
-	}
-}
-
-void Podium::serializeItemAttributes_OTBM(const IOMap& maphandle, NodeFileWriteHandle& stream) const {
-	Item::serializeItemAttributes_OTBM(maphandle, stream);
-
-	uint8_t flags = PODIUM_SHOW_OUTFIT * static_cast<uint8_t>(showOutfit) + PODIUM_SHOW_MOUNT * static_cast<uint8_t>(showMount) + PODIUM_SHOW_PLATFORM * static_cast<uint8_t>(showPlatform);
-
-	stream.addByte(OTBM_ATTR_PODIUMOUTFIT);
-	stream.addU8(flags);
-	stream.addU8(direction);
-
-	stream.addU16(outfit.lookType);
-	stream.addU8(outfit.lookHead);
-	stream.addU8(outfit.lookBody);
-	stream.addU8(outfit.lookLegs);
-	stream.addU8(outfit.lookFeet);
-	stream.addU8(outfit.lookAddon);
-
-	stream.addU16(outfit.lookMount);
-	stream.addU8(outfit.lookMountHead);
-	stream.addU8(outfit.lookMountBody);
-	stream.addU8(outfit.lookMountLegs);
-	stream.addU8(outfit.lookMountFeet);
-}
-
-/*
-	OTBM_ROOTV1
-	|
-	|--- OTBM_MAP_DATA
-	|	|
-	|	|--- OTBM_TILE_AREA
-	|	|	|--- OTBM_TILE
-	|	|	|--- OTBM_TILE_SQUARE (not implemented)
-	|	|	|--- OTBM_TILE_REF (not implemented)
-	|	|	|--- OTBM_HOUSETILE
-	|	|
-	|	|--- OTBM_SPAWNS (not implemented)
-	|	|	|--- OTBM_SPAWN_AREA (not implemented)
-	|	|	|--- OTBM_MONSTER (not implemented)
-	|	|
-	|	|--- OTBM_TOWNS
-	|		|--- OTBM_TOWN
-	|
-	|--- OTBM_ITEM_DEF (not implemented)
-*/
-
 bool IOMapOTBM::getVersionInfo(const FileName& filename, MapVersion& out_ver) {
 #ifdef OTGZ_SUPPORT
 	if (filename.GetExt() == "otgz") {
@@ -1096,257 +618,6 @@ bool IOMapOTBM::loadMap(Map& map, NodeFileReadHandle& f) {
 	return true;
 }
 
-bool IOMapOTBM::loadSpawns(Map& map, const FileName& dir) {
-	std::string fn = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvUTF8));
-	fn += map.spawnfile;
-
-	FileName filename(wxstr(fn));
-	if (!filename.FileExists()) {
-		warnings.push_back("IOMapOTBM::loadSpawns: File not found.");
-		return false;
-	}
-
-	// has to be declared again as encoding-specific characters break loading there
-	std::string encoded_path = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvWhateverWorks));
-	encoded_path += map.spawnfile;
-	pugi::xml_document doc;
-	pugi::xml_parse_result result = doc.load_file(encoded_path.c_str());
-	if (!result) {
-		warnings.push_back("IOMapOTBM::loadSpawns: File loading error.");
-		return false;
-	}
-	return loadSpawns(map, doc);
-}
-
-bool IOMapOTBM::loadSpawns(Map& map, pugi::xml_document& doc) {
-	pugi::xml_node node = doc.child("spawns");
-	if (!node) {
-		warnings.push_back("IOMapOTBM::loadSpawns: Invalid rootheader.");
-		return false;
-	}
-
-	for (pugi::xml_node spawnNode = node.first_child(); spawnNode; spawnNode = spawnNode.next_sibling()) {
-		if (as_lower_str(spawnNode.name()) != "spawn") {
-			continue;
-		}
-
-		Position spawnPosition;
-		spawnPosition.x = spawnNode.attribute("centerx").as_int();
-		spawnPosition.y = spawnNode.attribute("centery").as_int();
-		spawnPosition.z = spawnNode.attribute("centerz").as_int();
-
-		if (spawnPosition.x == 0 || spawnPosition.y == 0) {
-			warning("Bad position data on one spawn, discarding...");
-			continue;
-		}
-
-		int32_t radius = spawnNode.attribute("radius").as_int();
-		if (radius < 1) {
-			warning("Couldn't read radius of spawn.. discarding spawn...");
-			continue;
-		}
-
-		Tile* tile = map.getTile(spawnPosition);
-		if (tile && tile->spawn) {
-			warning("Duplicate spawn on position %d:%d:%d\n", tile->getX(), tile->getY(), tile->getZ());
-			continue;
-		}
-
-		Spawn* spawn = newd Spawn(radius);
-		if (!tile) {
-			tile = map.allocator(map.createTileL(spawnPosition));
-			map.setTile(spawnPosition, tile);
-		}
-
-		tile->spawn = spawn;
-		map.addSpawn(tile);
-
-		for (pugi::xml_node creatureNode = spawnNode.first_child(); creatureNode; creatureNode = creatureNode.next_sibling()) {
-			const std::string& creatureNodeName = as_lower_str(creatureNode.name());
-			if (creatureNodeName != "monster" && creatureNodeName != "npc") {
-				continue;
-			}
-
-			bool isNpc = creatureNodeName == "npc";
-			const std::string& name = creatureNode.attribute("name").as_string();
-			if (name.empty()) {
-				wxString err;
-				err << "Bad creature position data, discarding creature at spawn " << spawnPosition.x << ":" << spawnPosition.y << ":" << spawnPosition.z << " due missing name.";
-				warnings.Add(err);
-				break;
-			}
-
-			int32_t spawntime = creatureNode.attribute("spawntime").as_int();
-			if (spawntime == 0) {
-				spawntime = g_settings.getInteger(Config::DEFAULT_SPAWNTIME);
-			}
-
-			Direction direction = SOUTH;
-			int dir = creatureNode.attribute("direction").as_int(-1);
-			if (dir >= DIRECTION_FIRST && dir <= DIRECTION_LAST) {
-				direction = (Direction)dir;
-			}
-
-			Position creaturePosition(spawnPosition);
-
-			pugi::xml_attribute xAttribute = creatureNode.attribute("x");
-			pugi::xml_attribute yAttribute = creatureNode.attribute("y");
-			if (!xAttribute || !yAttribute) {
-				wxString err;
-				err << "Bad creature position data, discarding creature \"" << name << "\" at spawn " << creaturePosition.x << ":" << creaturePosition.y << ":" << creaturePosition.z << " due to invalid position.";
-				warnings.Add(err);
-				break;
-			}
-
-			creaturePosition.x += xAttribute.as_int();
-			creaturePosition.y += yAttribute.as_int();
-
-			radius = std::max<int32_t>(radius, std::abs(creaturePosition.x - spawnPosition.x));
-			radius = std::max<int32_t>(radius, std::abs(creaturePosition.y - spawnPosition.y));
-			radius = std::min<int32_t>(radius, g_settings.getInteger(Config::MAX_SPAWN_RADIUS));
-
-			Tile* creatureTile;
-			if (creaturePosition == spawnPosition) {
-				creatureTile = tile;
-			} else {
-				creatureTile = map.getTile(creaturePosition);
-			}
-
-			if (!creatureTile) {
-				wxString err;
-				err << "Discarding creature \"" << name << "\" at " << creaturePosition.x << ":" << creaturePosition.y << ":" << creaturePosition.z << " due to invalid position.";
-				warnings.Add(err);
-				break;
-			}
-
-			if (creatureTile->creature) {
-				wxString err;
-				err << "Duplicate creature \"" << name << "\" at " << creaturePosition.x << ":" << creaturePosition.y << ":" << creaturePosition.z << " was discarded.";
-				warnings.Add(err);
-				break;
-			}
-
-			CreatureType* type = g_creatures[name];
-			if (!type) {
-				type = g_creatures.addMissingCreatureType(name, isNpc);
-			} else {
-				if (type->outfit.lookType == 130) {
-					spdlog::info("Found existing creature in map: '{}' with lookType 130", name);
-				}
-			}
-
-			Creature* creature = newd Creature(type);
-			creature->setDirection(direction);
-			creature->setSpawnTime(spawntime);
-			creatureTile->creature = creature;
-
-			if (creatureTile->getLocation()->getSpawnCount() == 0) {
-				// No spawn, create a newd one
-				ASSERT(creatureTile->spawn == nullptr);
-				Spawn* spawn = newd Spawn(5);
-				creatureTile->spawn = spawn;
-				map.addSpawn(creatureTile);
-			}
-		}
-	}
-	return true;
-}
-
-bool IOMapOTBM::loadHouses(Map& map, const FileName& dir) {
-	std::string fn = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvUTF8));
-	fn += map.housefile;
-
-	FileName filename(wxstr(fn));
-	if (!filename.FileExists()) {
-		warnings.push_back("IOMapOTBM::loadHouses: File not found.");
-		return false;
-	}
-
-	// has to be declared again as encoding-specific characters break loading there
-	std::string encoded_path = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvWhateverWorks));
-	encoded_path += map.housefile;
-	pugi::xml_document doc;
-	pugi::xml_parse_result result = doc.load_file(encoded_path.c_str());
-	if (!result) {
-		warnings.push_back("IOMapOTBM::loadHouses: File loading error.");
-		return false;
-	}
-	return loadHouses(map, doc);
-}
-
-bool IOMapOTBM::loadHouses(Map& map, pugi::xml_document& doc) {
-	pugi::xml_node node = doc.child("houses");
-	if (!node) {
-		warnings.push_back("IOMapOTBM::loadHouses: Invalid rootheader.");
-		return false;
-	}
-
-	pugi::xml_attribute attribute;
-	for (pugi::xml_node houseNode = node.first_child(); houseNode; houseNode = houseNode.next_sibling()) {
-		if (as_lower_str(houseNode.name()) != "house") {
-			continue;
-		}
-
-		House* house = nullptr;
-		if ((attribute = houseNode.attribute("houseid"))) {
-			house = map.houses.getHouse(attribute.as_uint());
-			if (!house) {
-				break;
-			}
-		}
-
-		if ((attribute = houseNode.attribute("name"))) {
-			house->name = attribute.as_string();
-		} else {
-			house->name = "House #" + std::to_string(house->getID());
-		}
-
-		Position exitPosition(
-			houseNode.attribute("entryx").as_int(),
-			houseNode.attribute("entryy").as_int(),
-			houseNode.attribute("entryz").as_int()
-		);
-		if (exitPosition.x != 0 && exitPosition.y != 0 && exitPosition.z != 0) {
-			house->setExit(exitPosition);
-		}
-
-		if ((attribute = houseNode.attribute("rent"))) {
-			house->rent = attribute.as_int();
-		}
-
-		if ((attribute = houseNode.attribute("guildhall"))) {
-			house->guildhall = attribute.as_bool();
-		}
-
-		if ((attribute = houseNode.attribute("townid"))) {
-			house->townid = attribute.as_uint();
-		} else {
-			warning("House %d has no town! House was removed.", house->getID());
-			map.houses.removeHouse(house);
-		}
-	}
-	return true;
-}
-
-bool IOMapOTBM::loadWaypoints(Map& map, const FileName& dir) {
-	std::string fn = (const char*)(dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME).mb_str(wxConvUTF8));
-	fn += map.waypointfile;
-	FileName filename(wxstr(fn));
-	if (!filename.FileExists()) {
-		return false;
-	}
-
-	pugi::xml_document doc;
-	pugi::xml_parse_result result = doc.load_file(fn.c_str());
-	if (!result) {
-		return false;
-	}
-	return loadWaypoints(map, doc);
-};
-bool IOMapOTBM::loadWaypoints(Map& map, pugi::xml_document& doc) {
-	return true;
-};
-
 bool IOMapOTBM::saveMap(Map& map, const FileName& identifier) {
 #ifdef OTGZ_SUPPORT
 	if (identifier.GetExt() == "otgz") {
@@ -1673,158 +944,3 @@ bool IOMapOTBM::saveMap(Map& map, NodeFileWriteHandle& f) {
 	return true;
 }
 
-bool IOMapOTBM::saveSpawns(Map& map, const FileName& dir) {
-	wxString filepath = dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME);
-	filepath += wxString(map.spawnfile.c_str(), wxConvUTF8);
-
-	// Create the XML file
-	pugi::xml_document doc;
-	if (saveSpawns(map, doc)) {
-		return doc.save_file(filepath.wc_str(), "\t", pugi::format_default, pugi::encoding_utf8);
-	}
-	return false;
-}
-
-bool IOMapOTBM::saveSpawns(Map& map, pugi::xml_document& doc) {
-	pugi::xml_node decl = doc.prepend_child(pugi::node_declaration);
-	if (!decl) {
-		return false;
-	}
-
-	decl.append_attribute("version") = "1.0";
-
-	CreatureList creatureList;
-
-	pugi::xml_node spawnNodes = doc.append_child("spawns");
-	for (const auto& spawnPosition : map.spawns) {
-		Tile* tile = map.getTile(spawnPosition);
-		if (tile == nullptr) {
-			continue;
-		}
-
-		Spawn* spawn = tile->spawn;
-		ASSERT(spawn);
-
-		pugi::xml_node spawnNode = spawnNodes.append_child("spawn");
-
-		spawnNode.append_attribute("centerx") = spawnPosition.x;
-		spawnNode.append_attribute("centery") = spawnPosition.y;
-		spawnNode.append_attribute("centerz") = spawnPosition.z;
-
-		int32_t radius = spawn->getSize();
-		spawnNode.append_attribute("radius") = radius;
-
-		for (int32_t y = -radius; y <= radius; ++y) {
-			for (int32_t x = -radius; x <= radius; ++x) {
-				Tile* creature_tile = map.getTile(spawnPosition + Position(x, y, 0));
-				if (creature_tile) {
-					Creature* creature = creature_tile->creature;
-					if (creature && !creature->isSaved()) {
-						pugi::xml_node creatureNode = spawnNode.append_child(creature->isNpc() ? "npc" : "monster");
-
-						creatureNode.append_attribute("name") = creature->getName().c_str();
-						creatureNode.append_attribute("x") = x;
-						creatureNode.append_attribute("y") = y;
-						creatureNode.append_attribute("z") = spawnPosition.z;
-						creatureNode.append_attribute("spawntime") = creature->getSpawnTime();
-						if (creature->getDirection() != NORTH) {
-							creatureNode.append_attribute("direction") = creature->getDirection();
-						}
-
-						// Mark as saved
-						creature->save();
-						creatureList.push_back(creature);
-					}
-				}
-			}
-		}
-	}
-
-	for (Creature* creature : creatureList) {
-		creature->reset();
-	}
-	return true;
-}
-
-bool IOMapOTBM::saveHouses(Map& map, const FileName& dir) {
-	wxString filepath = dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME);
-	filepath += wxString(map.housefile.c_str(), wxConvUTF8);
-
-	// Create the XML file
-	pugi::xml_document doc;
-	if (saveHouses(map, doc)) {
-		return doc.save_file(filepath.wc_str(), "\t", pugi::format_default, pugi::encoding_utf8);
-	}
-	return false;
-}
-
-bool IOMapOTBM::saveHouses(Map& map, pugi::xml_document& doc) {
-	pugi::xml_node decl = doc.prepend_child(pugi::node_declaration);
-	if (!decl) {
-		return false;
-	}
-
-	decl.append_attribute("version") = "1.0";
-
-	pugi::xml_node houseNodes = doc.append_child("houses");
-	for (const auto& houseEntry : map.houses) {
-		const House* house = houseEntry.second;
-		pugi::xml_node houseNode = houseNodes.append_child("house");
-
-		houseNode.append_attribute("name") = house->name.c_str();
-		houseNode.append_attribute("houseid") = house->getID();
-
-		const Position& exitPosition = house->getExit();
-		houseNode.append_attribute("entryx") = exitPosition.x;
-		houseNode.append_attribute("entryy") = exitPosition.y;
-		houseNode.append_attribute("entryz") = exitPosition.z;
-
-		houseNode.append_attribute("rent") = house->rent;
-		if (house->guildhall) {
-			houseNode.append_attribute("guildhall") = true;
-		}
-
-		houseNode.append_attribute("townid") = house->townid;
-		houseNode.append_attribute("size") = static_cast<int32_t>(house->size());
-	}
-	return true;
-}
-
-bool IOMapOTBM::saveWaypoints(Map& map, const FileName& dir) {
-	wxString filepath = dir.GetPath(wxPATH_GET_SEPARATOR | wxPATH_GET_VOLUME);
-	filepath += wxString(map.waypointfile.c_str(), wxConvUTF8);
-
-	// Create the XML file
-	pugi::xml_document doc;
-	if (saveWaypoints(map, doc)) {
-		return doc.save_file(filepath.wc_str(), "\t", pugi::format_default, pugi::encoding_utf8);
-	}
-	return false;
-}
-
-bool IOMapOTBM::saveWaypoints(Map& map, pugi::xml_document& doc) {
-	pugi::xml_node decl = doc.prepend_child(pugi::node_declaration);
-	if (!decl) {
-		return false;
-	}
-
-	decl.append_attribute("version") = "1.0";
-
-	pugi::xml_node houseNodes = doc.append_child("waypoints");
-	for (const auto& houseEntry : map.houses) {
-		const House* house = houseEntry.second;
-		pugi::xml_node houseNode = houseNodes.append_child("waypoint");
-
-		houseNode.append_attribute("name") = house->name.c_str();
-		houseNode.append_attribute("id") = house->getID();
-		houseNode.append_attribute("icon") = house->getID();
-
-		const Position& exitPosition = house->getExit();
-		houseNode.append_attribute("x") = exitPosition.x;
-		houseNode.append_attribute("y") = exitPosition.y;
-		houseNode.append_attribute("z") = exitPosition.z;
-
-		houseNode.append_attribute("townid") = house->townid;
-	}
-	return true;
-}
